fix : 쿠폰 및 가격계산함수 분리
