+++ src/basic/utils/calculateCartTotal.ts
@@ -1,0 +1,38 @@
+import { CartItem, Coupon } from "../../types";
+import { calculateItemTotal } from "./calculateItemTotal";
+
+export const calculateCartTotal = (
+  cart: CartItem[],
+  selectedCoupon: Coupon | null
+) => {
+  let totalBeforeDiscount = 0;
+  let totalAfterDiscount = 0;
+
+  for (const item of cart) {
+    const itemPrice = item.product.price * item.quantity;
+    totalBeforeDiscount += itemPrice;
+    totalAfterDiscount += calculateItemTotal(item, cart);
+  }
+
+  if (selectedCoupon) {
+    if (selectedCoupon.discountType === "amount") {
+      totalAfterDiscount = Math.max(
+        0,
+        totalAfterDiscount - selectedCoupon.discountValue
+      );
+    } else {
+      totalAfterDiscount = Math.round(
+        totalAfterDiscount * (1 - selectedCoupon.discountValue / 100)
+      );
+    }
+  }
+
+  return {
+    totalBeforeDiscount: Math.round(totalBeforeDiscount),
+    totalAfterDiscount: Math.round(totalAfterDiscount),
+    discountAmount: Math.max(
+      0,
+      Math.round(totalBeforeDiscount - totalAfterDiscount)
+    ),
+  };
+};
